Changing floats --> doubles

+++ AssistantFunctions.h
@@ -43,7 +43,7 @@ void printMultiUMap(const unordered_multimap<string, string> &userTweetsRelation
 void printTweetsUmap(const unordered_map<string, Tweet> &tweets_umap);
 
 void ReadVaderLexicon_saveIt(const string &inVaderFileName , unordered_map<string ,double> &VaderUmap);
-void print_map (const unordered_map<string ,float> &VaderUmap);
+void print_map (const unordered_map<string ,double> &VaderUmap);
 
 void ReadCoinsFile_saveIt(const string &inCoinsFileName , vector<string> &CoinsList , unordered_map<string ,int> &CoinsUmap ); //todo thelei 5h sthlh kai opou den yparxei prwth
 void print_map(const unordered_map<string, int> &CoinsUmap) ;
